1_google_functions/2_insert_fetched_owners/requirements.txt
functions-framework==3.*

google-cloud-secret-manager